RavenDB-23250 fix review comments + update the about text
